+++ EngineMultiply.cc
@@ -9,16 +9,16 @@ static EngineRegistrationHelper<EngineMultiply>
                      { EngineFactory::ENGINE_IN_OPT_FILE_LIST});
 
 
-double EngineMultiply::run()
+uint64_t EngineMultiply::run()
 {
-  double result;
+  uint64_t result;
 
   if (!_mPOperandStream->getNext(result))
   {
     throw std::runtime_error("Null input");
   }
 
-  double operand;
+  uint64_t operand;
 
   while (_mPOperandStream->getNext(operand))
   {
